+++ scripts/upgrade-install-plugin.sh
@@ -5,13 +5,9 @@
 TARGET_PLUGIN="$1"
 TARGET_VERSION="$2"
 TARGET_PATH="$3"
+REPO_VERSION="$4"
 
 TARGET_PLUGIN_BUNDLE_NAME="${TARGET_PLUGIN/\//@}-${TARGET_VERSION}"
-
-REPO_VERSION="v1.0.0"
-if [[ "${TARGET_PLUGIN_BUNDLE_NAME:0:1}" == "@" ]]; then
-  REPO_VERSION="v1.0.0-1"
-fi
 
 REPO_BASE_URL="https://github.com/homebridge/plugins/releases/download/${REPO_VERSION}"